fix: fixed input forms to take names as mandatory and trow an error if they do not have a name
It has come to our attention that FormData class does not collect input data from any nameless input. This is why we made all of our inputs have names. If name is not used the error will be trown in the runtime in the browser askign you to use name for a given input with the ID declared. Make sure to have named ID's so that it is easier to troubleshoot which  element component is nameless.

Closes #40